# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hiring Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /positions:
    get:
      operationId: listPositions
      tags: [positions]
      summary: List all positions
      responses:
        "200":
          description: List of positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Position"
    post:
      operationId: createPosition
      tags: [positions]
      summary: Create a position
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePositionBody"
      responses:
        "201":
          description: Created position
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
  /positions/{id}:
    get:
      operationId: getPosition
      tags: [positions]
      summary: Get a position by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Position
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
        "404":
          description: Not found
    patch:
      operationId: updatePosition
      tags: [positions]
      summary: Update a position
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePositionBody"
      responses:
        "200":
          description: Updated position
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
    delete:
      operationId: deletePosition
      tags: [positions]
      summary: Delete a position
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /interviews:
    get:
      operationId: listInterviews
      tags: [interviews]
      summary: List all interviews
      parameters:
        - name: candidateId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of interviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Interview"
    post:
      operationId: createInterview
      tags: [interviews]
      summary: Schedule an interview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInterviewBody"
      responses:
        "201":
          description: Created interview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
  /interviews/{id}:
    get:
      operationId: getInterview
      tags: [interviews]
      summary: Get an interview by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Interview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
        "404":
          description: Not found
    patch:
      operationId: updateInterview
      tags: [interviews]
      summary: Update an interview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInterviewBody"
      responses:
        "200":
          description: Updated interview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
    delete:
      operationId: deleteInterview
      tags: [interviews]
      summary: Delete an interview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Position:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        department:
          type: string
        location:
          type: string
        type:
          type: string
          description: "full-time, part-time, contract, internship"
        status:
          type: string
          description: "open, paused, closed"
        description:
          type: ["string", "null"]
        headcount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - department
        - location
        - type
        - status
        - headcount
        - createdAt
        - updatedAt
    CreatePositionBody:
      type: object
      properties:
        title:
          type: string
        department:
          type: string
        location:
          type: string
        type:
          type: string
        status:
          type: string
        description:
          type: ["string", "null"]
        headcount:
          type: integer
      required:
        - title
        - department
        - location
        - type
        - status
        - headcount
    UpdatePositionBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        headcount:
          type: ["integer", "null"]
...
    Interview:
      type: object
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        candidateName:
          type: ["string", "null"]
        positionTitle:
          type: ["string", "null"]
        type:
          type: string
          description: "screening, technical, hr_interview, final_interview, case_study"
        scheduledAt:
          type: ["string", "null"]
          format: date-time
        durationMinutes:
          type: ["integer", "null"]
        interviewerName:
          type: ["string", "null"]
        interviewerEmail:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
          description: "video_call, phone, on_site"
        meetingUrl:
          type: ["string", "null"]
        status:
          type: string
          description: "scheduled, completed, cancelled, no_show"
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - candidateId
        - type
        - status
        - createdAt
    CreateInterviewBody:
      type: object
      properties:
        candidateId:
          type: integer
        type:
          type: string
        scheduledAt:
          type: ["string", "null"]
        durationMinutes:
          type: ["integer", "null"]
        interviewerName:
          type: ["string", "null"]
        interviewerEmail:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        meetingUrl:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - candidateId
        - type
        - status
    UpdateInterviewBody:
      type: object
      properties:
        type:
          type: ["string", "null"]
        scheduledAt:
          type: ["string", "null"]
        durationMinutes:
          type: ["integer", "null"]
        interviewerName:
          type: ["string", "null"]
        interviewerEmail:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        meetingUrl:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...