# lib/api-zod/src/generated/types/updatePositionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hiring Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdatePositionBody {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  headcount?: number | null;
}